Tensor Flow GPU Check and Test
This script contains checks to see that your GPU is visible to tensorflow and then two tests with tensorflow. The first compares TF to Torch in a basic operation. And the second does a basic image classification CNN training to compare GPU to CPU performance, this helps verify that the GPU is performing as expected (that's to say better than the CPU)